moved heating/cooling and tick handler to MPA-Lib

+++ src/main/java/com/github/lehjr/modularpowerarmor/client/gui/obsolete/ScrollableSlider.java
@@ -12,11 +12,11 @@ public class ScrollableSlider extends ScrollableRectangle {
         this.slider = slider;
     }
 
-    public float getValue() {
+    public double getValue() {
         return slider.getValue();
     }
 
-    public void setValue(float value) {
+    public void setValue(double value) {
         slider.setValue(value);
     }
